Restructure UI components related to the settings page
All "settings page" related stuff is moved into `ui/settings`. It is
also more modular now (settings page consists of multiple sections). It
should be fairly easy to add another "section" to the settings.
