+++ src/i18n/locales/en.json
@@ -76,7 +76,7 @@
   "upload-settings-modal-title": "Opencast Upload Settings",
   "upload-settings-button-store": "Save settings",
   "upload-settings-validation-error": "Invalid settings. Please verify and retry!",
-  "upload-settings-modal-header": "Configure Upload To Opencast",
+  "upload-settings-modal-header": "Connection to Opencast",
   "upload-settings-label-server-url": "Server URL",
   "upload-settings-label-workflow-id": "Workflow ID",
   "upload-settings-label-username": "Username",
